moving fftmorphology to own file in filters
